SFR-2678: Set record state
