+++ etl-pipeline/tests/functional/processes/ingest/test_doab_process.py
@@ -8,9 +8,10 @@ def test_doab_process(mock_epub_to_webpub):
     doab_process = DOABProcess('complete', None, None, None, 1, None)
     number_of_records_ingested = doab_process.runProcess()
 
+    records = assert_ingested_records(sources=[Source.DOAB.value], expected_number_of_records=number_of_records_ingested)
+
     record_pipeline_process = RecordPipelineProcess()
     record_pipeline_process.runProcess(max_attempts=2)
 
-    records = assert_ingested_records(sources=[Source.DOAB.value], expected_number_of_records=number_of_records_ingested)
     assert_uploaded_manifests(records)
     assert_uploaded_epubs(records)
