+++ etl-pipeline/tests/functional/processes/ingest/test_hathi_trust_process.py
@@ -8,8 +8,9 @@ def test_hathi_trust_process(mock_epub_to_webpub):
     hathi_trust_process = HathiTrustProcess('weekly', None, None, None, 5, None)
     number_of_records_ingested = hathi_trust_process.runProcess()
 
+    records = assert_ingested_records(sources=[Source.HATHI.value], expected_number_of_records=number_of_records_ingested)
+
     record_pipeline_process = RecordPipelineProcess()
     record_pipeline_process.runProcess(max_attempts=2)
 
-    records = assert_ingested_records(sources=[Source.HATHI.value], expected_number_of_records=number_of_records_ingested)
     assert_uploaded_manifests(records)
