+++ etl-pipeline/tests/functional/processes/ingest/test_loc_process.py
@@ -9,9 +9,10 @@ def test_loc_process(mock_epub_to_webpub):
     loc_process = LOCProcess('complete', None, None, None, 5, None)
     number_of_records_ingested = loc_process.runProcess()
 
+    records = assert_ingested_records(sources=[Source.LOC.value], expected_number_of_records=number_of_records_ingested)
+
     record_pipeline_process = RecordPipelineProcess()
     record_pipeline_process.runProcess(max_attempts=2)
 
-    records = assert_ingested_records(sources=[Source.LOC.value], expected_number_of_records=number_of_records_ingested)
     assert_uploaded_manifests(records)
     assert_uploaded_epubs(records)
