+++ etl-pipeline/tests/functional/processes/ingest/test_muse_process.py
@@ -9,9 +9,9 @@ def test_muse_process(mock_epub_to_webpub):
     muse_process = MUSEProcess('complete', None, None, None, 5, None)
     number_of_records_ingested = muse_process.runProcess()
 
+    records = assert_ingested_records(sources=[Source.MUSE.value], expected_number_of_records=number_of_records_ingested)
+
     record_pipeline_process = RecordPipelineProcess()
     record_pipeline_process.runProcess(max_attempts=2)
-
-    records = assert_ingested_records(sources=[Source.MUSE.value], expected_number_of_records=number_of_records_ingested)
 
     assert_uploaded_files(records)
